pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "dependency-injector>=4.48.2",
    "fastapi[standard]>=0.119.1",
    "greenlet>=3.2.4",
    "loguru>=0.7.3",
    "psutil>=7.1.1",
    "pydantic-settings>=2.11.0",
    "ruff>=0.14.3",
    "sqlalchemy>=2.0.44",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py313"

[tool.ruff.lint]
select = [
    "FAST",
    "ANN",
    "ASYNC",
    "S",
    "B",
    "A",
    "C4",
    "DTZ",
    "EM",
    "FIX",
    "FA",
    "ISC",
    "ICN",
    "LOG",
    "G",
    "PIE",
    "T20",
    "PYI",
    "PT",
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SIM",
    "TID",
    "TD",
    "TC",
    "ARG",
    "PTH",
    "FLY",
    "I",
    "C90",
    "N",
    "PERF",
    "E",
    "W",
    "F",
    "PGH",
    "PL",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "UP",
    "FURB",
    "RUF",
    "TRY"
]
ignore = [
    "FIX", # Delete this to detect pending work
    "TC003",
    "TC002",
    "TC001",
]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$" # Allow unused variables when underscore-prefixed.

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.commitizen]
name = "cz_gitmoji"
